Accelerometer FXLS8471Q java driver refined, now returned data is converted in int values representing acceleration from 4000mG to 4000mG

+++ VirtualSense/DJ_VirtualMachine.1.0/darjeeling.1.1/src/lib/virtualsense/javax/virtualsense/sensors/evm/accelerometer/AccVector.java
@@ -21,14 +21,14 @@
 package javax.virtualsense.sensors.evm.accelerometer ;
 
 /**
- * Acceleration vector.
+ * Acceleration vector. x,y,z acceleration components are specified in mG (range: -4000mG to 4000mG).
  * 
  * @author Matteo Dromedari
  *
  */
 public class AccVector
 {	
-	public short x;
-	public short y;
-	public short z;
+	public int x;
+	public int y;
+	public int z;
 }
